address comments and move GenerateKongRoutesFromGRPCRouteRule to translators
